+++ apps/sample-desktop/src/components/accountManagement/accountManagement.ts
@@ -1,5 +1,7 @@
 import type { IconName } from '@template/ui';
 
+export type TransferMenuState = 'withdraw' | 'transfer' | 'deposit';
+
 export interface TransferMenuItem {
   title: string;
   description: string;
